backend_flask/requirements.txt
alembic                  1.8.1
aniso8601                8.0.0
cachetools               5.2.0
certifi                  2022.9.24
charset-normalizer       2.1.1
click                    8.1.3
colorama                 0.4.5
Flask                    2.2.2
Flask-Cors               3.0.10
Flask-Migrate            3.1.0
Flask-RESTful            0.3.9
Flask-SQLAlchemy         2.5.1
google-api-core          2.10.1
google-api-python-client 2.63.0
google-auth              2.12.0
google-auth-httplib2     0.1.0
google-auth-oauthlib     0.5.3
googleapis-common-protos 1.56.4
httplib2                 0.20.4
idna                     3.4
importlib-metadata       4.12.0
itsdangerous             2.1.2
Jinja2                   3.1.2
Mako                     1.2.3
MarkupSafe               2.1.1
oauthlib                 3.2.1
pip                      22.2.2
protobuf                 4.21.6
psycopg2                 2.9.3
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pyparsing                3.0.9
python-dotenv            0.21.0
pytz                     2020.1
requests                 2.28.1
requests-oauthlib        1.3.1
rsa                      4.9
setuptools               65.4.1
six                      1.15.0
SQLAlchemy               1.3.18
tzdata                   2022.5
uritemplate              4.1.1
urllib3                  1.26.12
Werkzeug                 2.2.2
zipp                     3.8.1